+++ src/shared/components/palette/case-flag/components/select-flag-location/select-flag-location.component.spec.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormGroup, ReactiveFormsModule } from '@angular/forms';
-
+import { CaseFlagFieldState, SelectFlagLocationErrorMessage } from '../../enums';
 import { SelectFlagLocationComponent } from './select-flag-location.component';
 
 describe('SelectFlagLocationComponent', () => {
@@ -26,5 +26,25 @@ describe('SelectFlagLocationComponent', () => {
 
   it('should create component', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should set an error condition if the case has not been configured for flags', () => {
+    component.flagsData = [];
+    spyOn(component.caseFlagStateEmitter, 'emit');
+    component.ngOnInit();
+    expect(component.filteredFlagsData).toEqual([]);
+    expect(component.errorMessages[0]).toEqual({
+      title: '',
+      description: SelectFlagLocationErrorMessage.FLAGS_NOT_CONFIGURED,
+      fieldId: 'conditional-radios-list'
+    });
+    expect(component.caseFlagsConfigError).toBe(true);
+    expect(component.caseFlagStateEmitter.emit)
+      .toHaveBeenCalledWith({ currentCaseFlagFieldState: CaseFlagFieldState.FLAG_TYPE, errorMessages: component.errorMessages });
+    fixture.detectChanges();
+    const nativeElement = fixture.debugElement.nativeElement;
+    const nextButtonElement = nativeElement.querySelector('.button');
+    // The "Next" button should not be present if the error condition has been set
+    expect(nextButtonElement).toBeNull();
   });
 });