+++ src/shared/components/palette/case-flag/enums/select-flag-location-error-message.enum.ts
@@ -1,3 +1,4 @@
 export enum SelectFlagLocationErrorMessage {
-  FLAG_LOCATION_NOT_SELECTED = 'Please make a selection'
+  FLAG_LOCATION_NOT_SELECTED = 'Please make a selection',
+  FLAGS_NOT_CONFIGURED = 'Flags have not been configured for this case type'
 }